Created properties files with url, server and database info for four different environments - dev, uat, live, and local.  These are accessed using Spring's Environment class.  At the moment the correct properties file is selected at start up programmatically.  This is fine, or it can be removed and a System property can be set on the command line when the app is run.
Had to change the way the database connection is setup in the process - it
is now a Spring bean, defined in the EnvironmentSetup @Configuration
class, and injected into the EmployeeDAO class.

Future work: The properties files can be expanded, and a default
properties file can be added to remove duplication of properties that are
always the same.