Distribution lists are now generated in the end point manager/generateDistributionList and are built using one or both of the active directories, and then cached.  The cache is emptied  at the end of each day after the bulk update service has run.  This end point returns the distribution list and does not perform any other task, such as porposing an objective.
manager/proposeObjectiveToDistributionList
proposes an objective to distribution list that has previously been
generated and cached.  If the list is not in the cache, the objective will
not be proposed.

manager/proposeObjective now uses the DistributionList
interface.

Objective.isCurrent() altered so that it will never return false if the
progress is not complete.